Move trace directory
