+++ packages/next/server/dev/next-dev-server.ts
@@ -36,7 +36,7 @@ import { normalizePagePath } from '../normalize-page-path'
 import Router, { Params, route } from '../router'
 import { eventCliSession } from '../../telemetry/events'
 import { Telemetry } from '../../telemetry/storage'
-import { setGlobal } from '../../telemetry/trace'
+import { setGlobal } from '../../trace'
 import HotReloader from './hot-reloader'
 import { findPageFile } from '../lib/find-page-file'
 import { getNodeOptionsWithoutInspect } from '../lib/utils'
