+++ packages/next/trace/report/to-jaeger.ts
@@ -1,7 +1,7 @@
 import retry from 'next/dist/compiled/async-retry'
 import { randomBytes } from 'crypto'
 import fetch from 'node-fetch'
-import * as Log from '../../../build/output/log'
+import * as Log from '../../build/output/log'
 // Jaeger uses Zipkin's reporting
 import { batcher } from './to-zipkin'
