+++ packages/next/trace/report/to-json.ts
